requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
binaryornot==0.4.4
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
cloudpathlib==0.21.1
confection==0.1.5
cookiecutter==2.6.0
cryptography==44.0.2
cymem==2.0.11
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.0.1
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.1
django-phonenumber-field==8.1.0
django-tailwind==4.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-writable-nested==0.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
graphviz==0.20.3
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
odfpy==1.4.1
openai==1.84.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
phonenumbers==9.0.10
pillow==11.1.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydotplus==2.0.2
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
redis==6.2.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
twilio==9.6.5
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uuid==1.30
vine==5.1.0
Wand==0.6.13
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
